Extract BuildSettings to separate files
